Entity framework core migrations
PowerShellCommand
Get-Help about_entityframeworkcore
Add-Migration IntialMigration
Update-Database
It will create a micratmigration file.

+++ CoreWebApplication/Controllers/HomeController.cs
@@ -27,7 +27,7 @@ namespace CoreWebApplication.Controllers
         }
 
         [Route("{id?}")]
-        public ViewResult Details(int? id)
+        public ViewResult Details(long? id)
         {
             HomeDetailsViewModel homeDetailsViewModel = new HomeDetailsViewModel()
             {
